Move server files to the src folder (#2746)
